+++ css/style1.css
@@ -1,0 +1,71 @@
+*{
+    margin:0;
+    padding:0;
+    box-sizing: border-box;
+}
+
+body{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    background-color: #f7f7f7;
+    font-family: Arial, Helvetica, sans-serif;
+}
+.login-container{
+    background-color: #fff;
+    padding:30px;
+    border-radius: 10px;
+    box-shadow: 0 0 10px rgba(0,0,0,0.1);
+    width:100%;
+    max-width: 400px;
+    text-align: center;
+
+}
+.login-container h2{
+    margin-bottom: 20px;
+    font-size: 1.5rem;
+    color:#333;
+
+}
+
+.login-form .form-group{
+    margin-bottom: 15px;
+    text-align: left;
+
+}
+.login-form label{
+    display: block;
+    margin-bottom: 5px;
+    font-weight: bold;
+    color: #333;
+}
+
+.login-form input{
+    width:100%;
+    padding:10px;
+    border:1px solid #ddd;
+    border-radius: 5px;
+    font-size: 1rem;
+}
+
+.login-form input:focus{
+    border-color: #0066cc;
+    outline: none;
+
+}
+.login-button{
+    width: 100%;
+    padding: 10px;
+    background-color: #0066cc;
+    color: #fff;
+    border:none;
+    border-radius: 5px;
+    font-size: 1rem;
+    cursor: pointer;
+    transition: background-color 0.3s;
+}
+
+.login-button:hover{
+    background-color: #005bb5;
+}
