+++ html/product.html
@@ -1,0 +1,101 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+    <link rel="stylesheet" href="product.css">
+</head>
+
+<body>
+    <div class="main-title">
+        <div class="logo">
+            <img src="../img/ganpati-logo.jpg" alt="">
+        </div>
+        <div class="title-text">
+            <h1>Rajnish Company Ltd</h1>
+        </div>
+    </div>
+
+    <section class="pro-body">
+        <h2 class="heading">Latest <span>Products</span></h2>
+
+        <div class="pro-container">
+
+            <div class="pro-box">
+                <!-- <img src="../img/pro1.jpg" alt=""> -->
+
+                <h3>Electrial Material</h3>
+                <p>vhguyuiuiuuoiujiyuyjtdtrdytfygtyd/vhguyuiuiuuoiujiyuyjtdtrdytfygtyd>
+                </p>
+
+            </div>
+
+            <div class="pro-box">
+                <h3>Hardware</h3>
+                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Consectetur asperiores laboriosam fuga
+                    commodi
+                    molestiae! Perspiciatis vero a velit ipsum vitae dignissimos nemo saepe ad animi illum! Porro
+                    officia
+                    possimus in?</p>
+            </div>
+            <div class="pro-box">
+                <h3>Bathroom & Pipeline</h3>
+                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Consectetur asperiores laboriosam fuga
+                    commodi
+                    molestiae! Perspiciatis vero a velit ipsum vitae dignissimos nemo saepe ad animi illum! Porro
+                    officia
+                    possimus in?</p>
+            </div>
+
+            <div class="pro-box">
+                <h3>Contrete</h3>
+                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Consectetur asperiores laboriosam fuga
+                    commodi
+                    molestiae! Perspiciatis vero a velit ipsum vitae dignissimos nemo saepe ad animi illum! Porro
+                    officia
+                    possimus in?</p>
+            </div>
+
+            <div class="pro-box">
+                <h3>Steel</h3>
+                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Consectetur asperiores laboriosam fuga
+                    commodi
+                    molestiae! Perspiciatis vero a velit ipsum vitae dignissimos nemo saepe ad animi illum! Porro
+                    officia
+                    possimus in?</p>
+            </div>
+
+            <div class="pro-box">
+                <h3>Sand</h3>
+                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Consectetur asperiores laboriosam fuga
+                    commodi
+                    molestiae! Perspiciatis vero a velit ipsum vitae dignissimos nemo saepe ad animi illum! Porro
+                    officia
+                    possimus in?</p>
+            </div>
+
+            <div class="pro-box">
+                <h3>Cement</h3>
+                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Consectetur asperiores laboriosam fuga
+                    commodi
+                    molestiae! Perspiciatis vero a velit ipsum vitae dignissimos nemo saepe ad animi illum! Porro
+                    officia
+                    possimus in?</p>
+            </div>
+
+            <div class="pro-box">
+                <h3>Machinery for Construction & Maintenance</h3>
+                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Consectetur asperiores laboriosam fuga
+                    commodi
+                    molestiae! Perspiciatis vero a velit ipsum vitae dignissimos nemo saepe ad animi illum! Porro
+                    officia
+                    possimus in?</p>
+            </div>
+        </div>
+    </section>
+
+</body>
+
+</html>
